Ex12
The operating system. On this exerice we have built a few classes that enables the user to write in the Jack language with classes like:
"Screen", "Math", "Array" and "Memory", the user could write any high level code that the language offers, which would be translated to VM, which would be translated to asm and so on.